if else ugh

+++ config/setup.sh
@@ -5,10 +5,13 @@ apt-get -q -y install git libreoffice unattended-upgrades
 
 dpkg-reconfigure -plow unattended-upgrades
 
-git clone https://github.com/karlsander/pcr-guest-home /home/nbadmin/pcr-guest-home/
-cd /home/nbadmin/pcr-guest-home
-git fetch --all
-git reset --hard origin/master
+if [ -d "/home/nbadmin/pcr-guest-home/.git" ]; then
+    cd /home/nbadmin/pcr-guest-home
+    git fetch --all
+    git reset --hard origin/master
+else
+    git clone https://github.com/karlsander/pcr-guest-home /home/nbadmin/pcr-guest-home/
+fi
 
 rm -rf /etc/guest-session/
 mkdir /etc/guest-session/
